requirements.txt
langchain==0.3.13
langchain-community==0.3.13
langchain-groq==0.2.2
unstructured==0.11.8
selenium==4.27.1
streamlit==1.41.1
pandas==2.2.3
python-dotenv==1.0.1
sentence-transformers==2.2.2
pinecone-client==2.2.3
